Add stock changes (#7)
- Update Double to BigDecimal
- Add ask- and bid-price
- Remove unnecessary print statement
-Remove old methods that are no longer used
- Small changes

+++ src/main/java/com/scaleup/backend/league/DTO/LeagueDTO.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.util.LinkedHashMap;
 
 @Getter
@@ -15,8 +16,8 @@ public class LeagueDTO {
     private String leagueId;
     private String leagueName;
     private String code;
-    private Double startBudget;
-    private Double transactionCost;
+    private BigDecimal startBudget;
+    private BigDecimal transactionCost;
     private Integer stockAmount;
     private LinkedHashMap<String, Double> probability;
 }
